+++ src/PhpSpreadsheet/Comment.php
@@ -2,10 +2,13 @@
 
 namespace PhpOffice\PhpSpreadsheet;
 
+use PhpOffice\PhpSpreadsheet\Exception as PhpSpreadsheetException;
 use PhpOffice\PhpSpreadsheet\Helper\Size;
 use PhpOffice\PhpSpreadsheet\RichText\RichText;
+use PhpOffice\PhpSpreadsheet\Shared\Drawing as SharedDrawing;
 use PhpOffice\PhpSpreadsheet\Style\Alignment;
 use PhpOffice\PhpSpreadsheet\Style\Color;
+use PhpOffice\PhpSpreadsheet\Worksheet\Drawing;
 
 class Comment implements IComparable
 {
@@ -73,6 +76,13 @@ class Comment implements IComparable
     private $alignment;
 
     /**
+     * Background image in comment.
+     *
+     * @var Drawing
+     */
+    private $backgroundImage;
+
+    /**
      * Create a new Comment.
      */
     public function __construct()
@@ -82,6 +92,7 @@ class Comment implements IComparable
         $this->text = new RichText();
         $this->fillColor = new Color('FFFFFFE1');
         $this->alignment = Alignment::HORIZONTAL_GENERAL;
+        $this->backgroundImage = new Drawing();
     }
 
     /**
@@ -273,6 +284,7 @@ class Comment implements IComparable
             ($this->visible ? 1 : 0) .
             $this->fillColor->getHashCode() .
             $this->alignment .
+            ($this->hasBackgroundImage() ? $this->backgroundImage->getHashCode() : '') .
             __CLASS__
         );
     }
@@ -298,5 +310,53 @@ class Comment implements IComparable
     public function __toString(): string
     {
         return $this->text->getPlainText();
+    }
+
+    /**
+     * Check is background image exists.
+     */
+    public function hasBackgroundImage(): bool
+    {
+        $path = $this->backgroundImage->getPath();
+
+        if (empty($path)) {
+            return false;
+        }
+
+        return getimagesize($path) !== false;
+    }
+
+    /**
+     * Returns background image.
+     */
+    public function getBackgroundImage(): Drawing
+    {
+        return $this->backgroundImage;
+    }
+
+    /**
+     * Sets background image.
+     */
+    public function setBackgroundImage(Drawing $objDrawing): self
+    {
+        if (!array_key_exists($objDrawing->getType(), Drawing::IMAGE_TYPES_CONVERTION_MAP)) {
+            throw new PhpSpreadsheetException('Unsupported image type in comment background. Supported types: PNG, JPEG, BMP, GIF.');
+        }
+        $this->backgroundImage = $objDrawing;
+
+        return $this;
+    }
+
+    /**
+     * Sets size of comment as size of background image.
+     */
+    public function setSizeAsBackgroundImage(): self
+    {
+        if ($this->hasBackgroundImage()) {
+            $this->setWidth(SharedDrawing::pixelsToPoints($this->backgroundImage->getWidth()) . 'pt');
+            $this->setHeight(SharedDrawing::pixelsToPoints($this->backgroundImage->getHeight()) . 'pt');
+        }
+
+        return $this;
     }
 }