+++ src/PhpSpreadsheet/Worksheet/Drawing.php
@@ -3,9 +3,17 @@
 namespace PhpOffice\PhpSpreadsheet\Worksheet;
 
 use PhpOffice\PhpSpreadsheet\Exception as PhpSpreadsheetException;
+use ZipArchive;
 
 class Drawing extends BaseDrawing
 {
+    const IMAGE_TYPES_CONVERTION_MAP = [
+        IMAGETYPE_GIF => IMAGETYPE_PNG,
+        IMAGETYPE_JPEG => IMAGETYPE_JPEG,
+        IMAGETYPE_PNG => IMAGETYPE_PNG,
+        IMAGETYPE_BMP => IMAGETYPE_PNG,
+    ];
+
     /**
      * Path.
      *
@@ -48,10 +56,7 @@ class Drawing extends BaseDrawing
      */
     public function getIndexedFilename(): string
     {
-        $fileName = $this->getFilename();
-        $fileName = str_replace(' ', '_', $fileName);
-
-        return str_replace('.' . $this->getExtension(), '', $fileName) . $this->getImageIndex() . '.' . $this->getExtension();
+        return md5($this->path) . '.' . $this->getExtension();
     }
 
     /**
@@ -64,6 +69,20 @@ class Drawing extends BaseDrawing
         $exploded = explode('.', basename($this->path));
 
         return $exploded[count($exploded) - 1];
+    }
+
+    /**
+     * Get full filepath to store drawing in zip archive.
+     *
+     * @return string
+     */
+    public function getMediaFilename()
+    {
+        if (!array_key_exists($this->type, self::IMAGE_TYPES_CONVERTION_MAP)) {
+            throw new PhpSpreadsheetException('Unsupported image type in comment background. Supported types: PNG, JPEG, BMP, GIF.');
+        }
+
+        return sprintf('image%d%s', $this->getImageIndex(), $this->getImageFileExtensionForSave());
     }
 
     /**
@@ -81,10 +100,11 @@ class Drawing extends BaseDrawing
      *
      * @param string $path File path
      * @param bool $verifyFile Verify file
+     * @param ZipArchive $zip Zip archive instance
      *
      * @return $this
      */
-    public function setPath($path, $verifyFile = true)
+    public function setPath($path, $verifyFile = true, $zip = null)
     {
         if ($verifyFile) {
             // Check if a URL has been passed. https://stackoverflow.com/a/2058596/1252979
@@ -97,18 +117,18 @@ class Drawing extends BaseDrawing
                 if ($filePath) {
                     file_put_contents($filePath, $imageContents);
                     if (file_exists($filePath)) {
-                        if ($this->width == 0 && $this->height == 0) {
-                            // Get width/height
-                            [$this->width, $this->height] = getimagesize($filePath);
-                        }
+                        $this->setSizesAndType($filePath);
                         unlink($filePath);
                     }
                 }
             } elseif (file_exists($path)) {
                 $this->path = $path;
-                if ($this->width == 0 && $this->height == 0) {
-                    // Get width/height
-                    [$this->width, $this->height] = getimagesize($path);
+                $this->setSizesAndType($path);
+            } elseif ($zip instanceof ZipArchive) {
+                $zipPath = explode('#', $path)[1];
+                if ($zip->locateName($zipPath) !== false) {
+                    $this->path = $path;
+                    $this->setSizesAndType($path);
                 }
             } else {
                 throw new PhpSpreadsheetException("File $path not found!");
@@ -152,5 +172,43 @@ class Drawing extends BaseDrawing
             parent::getHashCode() .
             __CLASS__
         );
+    }
+
+    /**
+     * Get Image Type for Save.
+     */
+    public function getImageTypeForSave(): int
+    {
+        if (!array_key_exists($this->type, self::IMAGE_TYPES_CONVERTION_MAP)) {
+            throw new PhpSpreadsheetException('Unsupported image type in comment background. Supported types: PNG, JPEG, BMP, GIF.');
+        }
+
+        return self::IMAGE_TYPES_CONVERTION_MAP[$this->type];
+    }
+
+    /**
+     * Get Image file extention for Save.
+     */
+    public function getImageFileExtensionForSave(bool $includeDot = true): string
+    {
+        if (!array_key_exists($this->type, self::IMAGE_TYPES_CONVERTION_MAP)) {
+            throw new PhpSpreadsheetException('Unsupported image type in comment background. Supported types: PNG, JPEG, BMP, GIF.');
+        }
+
+        $result = image_type_to_extension(self::IMAGE_TYPES_CONVERTION_MAP[$this->type], $includeDot);
+
+        return is_string($result) ? $result : '';
+    }
+
+    /**
+     * Get Image mime type.
+     */
+    public function getImageMimeType(): string
+    {
+        if (!array_key_exists($this->type, self::IMAGE_TYPES_CONVERTION_MAP)) {
+            throw new PhpSpreadsheetException('Unsupported image type in comment background. Supported types: PNG, JPEG, BMP, GIF.');
+        }
+
+        return image_type_to_mime_type(self::IMAGE_TYPES_CONVERTION_MAP[$this->type]);
     }
 }